feat(refit): off-rail detection guards for --iterate
Three new safety checks in the iterate pipeline:

1. blocked_on_dirty_entry: fires when worktree has uncommitted changes
   at gate-entry in a fresh (non-retry) state. Without this, git add -A
   would silently bundle agent-authored changes into the gate-fix commit.

2. warn_config_drift: non-blocking warning when .sb_config.json changes
   after plan generation. Config hash stored at --start time; compared
   on every --iterate invocation. Gate list may be stale.

3. blocked_on_head_drift recovery hint: the else-branch message now
   includes "sm refit --skip" as the explicit escape hatch when an agent
   manually committed a fix outside the iterate rail.

7 new unit tests across TestDirtyEntryGuard, TestConfigDriftWarning,
TestHeadDriftRecoveryHint. TestCmdRefitContinue split to its own file
(test_refit_continue.py) to stay within code-sprawl limit. Shared
_FakeLock helper moved to tests/conftest.py.
